syncify try_start_burn and move it into simple_ui
